First messy stab at refactoring
