+++ cdk-api/package.json
@@ -38,6 +38,7 @@
     "@aws-sdk/client-ecr": "^3.229.0",
     "@aws-sdk/client-s3": "^3.180.0",
     "@aws-sdk/lib-dynamodb": "^3.234.0",
+    "@connected-web/openapi-rest-api": "^0.2.1",
     "ajv": "^8.12.0",
     "ajv-formats": "^2.1.1",
     "aws-cdk-lib": "2.80.0",
