+++ cdk-api/src/stack/endpoints/Status/handler.ts
@@ -2,7 +2,7 @@ import {
   APIGatewayProxyEvent,
   APIGatewayProxyResult
 } from 'aws-lambda/trigger/api-gateway-proxy'
-import { lambdaResponse } from './helpers/lambdaResponse'
+import { lambdaResponse } from '../helpers/lambdaResponse'
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   const statusInfo = process.env.STATUS_INFO ?? JSON.stringify({ message: 'No STATUS_INFO found on env' })
