+++ cdk-api/src/stack/endpoints/helpers/aws/deleteObject.ts
