+++ cdk-api/src/stack/endpoints/helpers/aws/getObject.ts
