+++ cdk-api/src/stack/endpoints/helpers/aws/listObjects.ts
