+++ cdk-api/src/stack/endpoints/helpers/aws/putObject.ts
