+++ cdk-api/src/stack/endpoints/helpers/aws/s3Client.ts
