+++ cdk-api/src/stack/endpoints/helpers/corsHeaders.ts
