+++ cdk-api/src/stack/endpoints/helpers/getJson.ts
