+++ cdk-api/src/stack/endpoints/helpers/interfaces.ts
