+++ cdk-api/src/stack/endpoints/helpers/lambdaResponse.ts
