+++ cdk-api/src/stack/endpoints/helpers/responses.ts
