+++ cdk-api/src/stack/endpoints/helpers/userScopes.ts
